refactor: use existing paypal commerce class to register with new gateway api
