+++ src/PaymentGateways/Actions/RegisterPaymentGateways.php
@@ -5,7 +5,7 @@ namespace Give\PaymentGateways\Actions;
 use Give\Framework\Exceptions\Primitives\Exception;
 use Give\Framework\Exceptions\Primitives\InvalidArgumentException;
 use Give\Framework\PaymentGateways\PaymentGatewayRegister;
-use Give\PaymentGateways\PayPalCommerce\PayPalCommerceGateway;
+use Give\PaymentGateways\PayPalCommerce\PayPalCommerce;
 
 class RegisterPaymentGateways
 {
@@ -18,7 +18,7 @@ class RegisterPaymentGateways
         // When complete, the Test Gateway will eventually replace The legacy Manual Gateway.
         //TestGateway::class,
         //TestGatewayOffsite::class,
-        PayPalCommerceGateway::class
+        PayPalCommerce::class
     ];
 
     /**
